Add plugin manager and update imports
