+++ tests/test_plugin_registry.py
@@ -6,7 +6,7 @@ from typing import Callable
 import pytest
 httpx = pytest.importorskip("httpx")
 
-import genecoder.plugins as plugins
+import genecoder.plugin_manager as plugins
 from genecoder.security import compute_checksum
 import base64
 from pathlib import Path
